Chapter24終了（デザインが反映されない）
